requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai>=1.3.0
PyPDF2>=3.0.0
python-docx>=0.8.11
numpy>=1.24.3
faiss-cpu>=1.7.4
Werkzeug>=3.0.0
requests>=2.31.0
